Rename server/pool.ts to server/runner.ts
